[CHA-2532] add required prop to allow closing the modal only with the modal buttons
